+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Updated to latest origins version (1.13.0-pre.2)
+- Updated to latest origins version (1.13.0-pre.3)